opt: refactor and move some functions for locality optimized search
This commit refactors some functions previously in xform/scan_funcs.go
for GenerateLocalityOptimizedScan and moves them to xform/general_funcs.go
so they can be used to generate locality optimized anti joins in a future
commit.

Release note: None
