+++ pkg/sql/opt/xform/general_funcs.go
@@ -11,13 +11,17 @@
 package xform
 
 import (
+	"sort"
+
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/cat"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/idxconstraint"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/norm"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/partialidx"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/errors"
 )
 
@@ -322,4 +326,151 @@ func (c *CustomFuncs) findConstantFilterCols(
 			}
 		}
 	}
+}
+
+// isZoneLocal returns true if the given zone config indicates that the replicas
+// it constrains will be primarily located in the localRegion.
+func isZoneLocal(zone cat.Zone, localRegion string) bool {
+	// First count the number of local and remote replica constraints. If all
+	// are local or all are remote, we can return early.
+	local, remote := 0, 0
+	for i, n := 0, zone.ReplicaConstraintsCount(); i < n; i++ {
+		replicaConstraint := zone.ReplicaConstraints(i)
+		for j, m := 0, replicaConstraint.ConstraintCount(); j < m; j++ {
+			constraint := replicaConstraint.Constraint(j)
+			if isLocal, ok := isConstraintLocal(constraint, localRegion); ok {
+				if isLocal {
+					local++
+				} else {
+					remote++
+				}
+			}
+		}
+	}
+	if local > 0 && remote == 0 {
+		return true
+	}
+	if remote > 0 && local == 0 {
+		return false
+	}
+
+	// Next check the voter replica constraints. Once again, if all are local or
+	// all are remote, we can return early.
+	local, remote = 0, 0
+	for i, n := 0, zone.VoterConstraintsCount(); i < n; i++ {
+		replicaConstraint := zone.VoterConstraint(i)
+		for j, m := 0, replicaConstraint.ConstraintCount(); j < m; j++ {
+			constraint := replicaConstraint.Constraint(j)
+			if isLocal, ok := isConstraintLocal(constraint, localRegion); ok {
+				if isLocal {
+					local++
+				} else {
+					remote++
+				}
+			}
+		}
+	}
+	if local > 0 && remote == 0 {
+		return true
+	}
+	if remote > 0 && local == 0 {
+		return false
+	}
+
+	// Use the lease preferences as a tie breaker. We only really care about the
+	// first one, since subsequent lease preferences only apply in edge cases.
+	if zone.LeasePreferenceCount() > 0 {
+		leasePref := zone.LeasePreference(0)
+		for i, n := 0, leasePref.ConstraintCount(); i < n; i++ {
+			constraint := leasePref.Constraint(i)
+			if isLocal, ok := isConstraintLocal(constraint, localRegion); ok {
+				return isLocal
+			}
+		}
+	}
+
+	return false
+}
+
+// isConstraintLocal returns isLocal=true and ok=true if the given constraint is
+// a required constraint matching the given localRegion. Returns isLocal=false
+// and ok=true if the given constraint is a prohibited constraint matching the
+// given local region or if it is a required constraint matching a different
+// region. Any other scenario returns ok=false, since this constraint gives no
+// information about whether the constrained replicas are local or remote.
+func isConstraintLocal(constraint cat.Constraint, localRegion string) (isLocal bool, ok bool) {
+	if constraint.GetKey() != regionKey {
+		// We only care about constraints on the region.
+		return false /* isLocal */, false /* ok */
+	}
+	if constraint.GetValue() == localRegion {
+		if constraint.IsRequired() {
+			// The local region is required.
+			return true /* isLocal */, true /* ok */
+		}
+		// The local region is prohibited.
+		return false /* isLocal */, true /* ok */
+	}
+	if constraint.IsRequired() {
+		// A remote region is required.
+		return false /* isLocal */, true /* ok */
+	}
+	// A remote region is prohibited, so this constraint gives no information
+	// about whether the constrained replicas are local or remote.
+	return false /* isLocal */, false /* ok */
+}
+
+// prefixIsLocal contains a PARTITION BY LIST prefix, and a boolean indicating
+// whether the prefix is from a local partition.
+type prefixIsLocal struct {
+	prefix  tree.Datums
+	isLocal bool
+}
+
+// prefixSorter sorts prefixes (which are wrapped in prefixIsLocal structs) so
+// that longer prefixes are ordered first.
+type prefixSorter []prefixIsLocal
+
+var _ sort.Interface = &prefixSorter{}
+
+// Len is part of sort.Interface.
+func (ps prefixSorter) Len() int {
+	return len(ps)
+}
+
+// Less is part of sort.Interface.
+func (ps prefixSorter) Less(i, j int) bool {
+	return len(ps[i].prefix) > len(ps[j].prefix)
+}
+
+// Swap is part of sort.Interface.
+func (ps prefixSorter) Swap(i, j int) {
+	ps[i], ps[j] = ps[j], ps[i]
+}
+
+// getSortedPrefixes collects all the prefixes from all the different partitions
+// in the index (remembering which ones came from local partitions), and sorts
+// them so that longer prefixes come before shorter prefixes.
+func getSortedPrefixes(index cat.Index, localPartitions util.FastIntSet) []prefixIsLocal {
+	allPrefixes := make(prefixSorter, 0, index.PartitionCount())
+	for i, n := 0, index.PartitionCount(); i < n; i++ {
+		part := index.Partition(i)
+		isLocal := localPartitions.Contains(i)
+		partitionPrefixes := part.PartitionByListPrefixes()
+		if len(partitionPrefixes) == 0 {
+			// This can happen when the partition value is DEFAULT.
+			allPrefixes = append(allPrefixes, prefixIsLocal{
+				prefix:  nil,
+				isLocal: isLocal,
+			})
+		}
+		for j := range partitionPrefixes {
+			allPrefixes = append(allPrefixes, prefixIsLocal{
+				prefix:  partitionPrefixes[j],
+				isLocal: isLocal,
+			})
+		}
+	}
+	sort.Sort(allPrefixes)
+	return allPrefixes
 }
