+++ pkg/sql/opt/xform/general_funcs_test.go
